# your_project/docs/swagger.yaml
basePath: /api
definitions:
  ApiResponse:
    description: Generic API response
    properties:
      data: {}
      responseKey:
        type: string
    type: object
...
  RolesCount:
    description: RolesCount represents the count of roles.
    properties:
      count:
        description: Count is the number of roles.
        type: integer
    type: object
  RolesDetails:
    description: RolesDetails represents detailed information about a specific role.
    properties:
      companyID:
        description: CompanyID is the unique identifier for the company associated
          with the role.
        type: string
      companyName:
        description: CompanyName is the name of the company associated with the role.
        type: string
      createdAt:
        description: CreatedAt is the timestamp indicating when the role entry was
          created.
        type: string
      firstname:
        description: Name is the name of the role.
        type: string
      id:
        description: ID is the unique identifier for the role.
        type: string
    type: object
  RolesIn:
    description: RolesIn represents the input structure for creating a new role.
    properties:
      name:
        description: Name is the name of the role. It is required and should be between
          2 and 40 characters.
        maxLength: 40
        minLength: 2
        type: string
    required:
    - name
    type: object
  RolesList:
    description: RolesList represents a simplified version of the role for listing
      purposes.
    properties:
      id:
        description: ID is the unique identifier for the role.
        type: string
      name:
        description: Name is the name of the role.
        type: string
    type: object
  RolesPagination:
    description: RolesPagination represents the paginated list of roles.
    properties:
      items:
        description: Items is a slice containing individual role details.
        items:
          $ref: '#/definitions/RolesTable'
        type: array
      limit:
        description: Limit is the maximum number of items per page in the pagination.
        type: integer
      page:
        description: Page is the current page number in the pagination.
        type: integer
      totalCount:
        description: TotalCount is the total number of roles in the entire list.
        type: integer
    type: object
  RolesTable:
    description: RolesTable represents a single role entry in a table.
    properties:
      createdAt:
        description: CreatedAt is the timestamp indicating when the role entry was
          created.
        type: string
      id:
        description: ID is the unique identifier for the role.
        type: string
      name:
        description: Name is the name of the role.
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: h.manai@tunisiancloud.com
    name: Hedi Manai
  description: A basic template for managing your project using Golang REST APIs.
  title: Project Name
  version: "0.1"
paths:
...
  /roles/{companyID}:
    get:
      description: Get all roles.
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RolesPagination'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Create a new role.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: Role query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/RolesIn'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Create role
      tags:
      - Roles
  /roles/{companyID}/{ID}:
    delete:
      description: Delete one role.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: Role ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete role
      tags:
      - Roles
    get:
      description: Get one role.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: Role ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RolesDetails'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get role
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Update one role.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: Role ID
        in: path
        name: ID
        required: true
        type: string
      - description: Role query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/RolesIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Update role
      tags:
      - Roles
  /roles/{companyID}/count:
    get:
      description: Get number of all roles.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RolesCount'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get number of  roles
      tags:
      - Roles
  /roles/{companyID}/list:
    get:
      description: Get list of all roles.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/RolesList'
            type: array
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get list of  roles
      tags:
      - Roles
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"